+++ src/script/compiler/ast/AstNil.hpp
@@ -15,8 +15,10 @@ public:
 
     virtual Tribool IsTrue() const override;
     virtual bool IsNumber() const override;
-    virtual hyperion::int32 IntValue() const override;
-    virtual float FloatValue() const override;
+
+    virtual hyperion::int64 IntValue() const override;
+    virtual double FloatValue() const override;
+
     virtual SymbolTypeRef GetExprType() const override;
 
     virtual RC<AstConstant> HandleOperator(Operators opType, const AstConstant* right) const override;
@@ -24,8 +26,7 @@ public:
 private:
     RC<AstNil> CloneImpl() const
     {
-        return RC<AstNil>(new AstNil(
-            m_location));
+        return RC<AstNil>(new AstNil(m_location));
     }
 };
