+++ src/script/compiler/ast/AstString.hpp
@@ -23,8 +23,10 @@ public:
 
     virtual Tribool IsTrue() const override;
     virtual bool IsNumber() const override;
-    virtual hyperion::int32 IntValue() const override;
-    virtual float FloatValue() const override;
+
+    virtual hyperion::int64 IntValue() const override;
+    virtual double FloatValue() const override;
+
     virtual SymbolTypeRef GetExprType() const override;
 
     virtual RC<AstConstant> HandleOperator(Operators opType, const AstConstant* right) const override;
