+++ src/script/compiler/ast/AstTrue.hpp
@@ -15,8 +15,10 @@ public:
 
     virtual Tribool IsTrue() const override;
     virtual bool IsNumber() const override;
-    virtual hyperion::int32 IntValue() const override;
-    virtual float FloatValue() const override;
+
+    virtual hyperion::int64 IntValue() const override;
+    virtual double FloatValue() const override;
+
     virtual SymbolTypeRef GetExprType() const override;
 
     virtual RC<AstConstant> HandleOperator(Operators opType, const AstConstant* right) const override;
@@ -34,8 +36,7 @@ public:
 private:
     RC<AstTrue> CloneImpl() const
     {
-        return RC<AstTrue>(new AstTrue(
-            m_location));
+        return RC<AstTrue>(new AstTrue(m_location));
     }
 };
