+++ src/script/compiler/ast/AstUndefined.hpp
@@ -15,8 +15,10 @@ public:
 
     virtual Tribool IsTrue() const override;
     virtual bool IsNumber() const override;
-    virtual hyperion::int32 IntValue() const override;
-    virtual float FloatValue() const override;
+
+    virtual hyperion::int64 IntValue() const override;
+    virtual double FloatValue() const override;
+
     virtual SymbolTypeRef GetExprType() const override;
 
     virtual RC<AstConstant> HandleOperator(Operators opType, const AstConstant* right) const override;
@@ -24,8 +26,7 @@ public:
 private:
     RC<AstUndefined> CloneImpl() const
     {
-        return RC<AstUndefined>(new AstUndefined(
-            m_location));
+        return RC<AstUndefined>(new AstUndefined(m_location));
     }
 };
